@@ -16,12 +16,12 @@
     }
   ],
   "require": {
-    "php": "^8.1",
+    "php": "^8.2",
     "laravel/framework": "^11.0 || ^12.0",
     "mrclay/minify": "^3.0"
   },
   "require-dev": {
-    "orchestra/testbench": "^9.0|^10.0",
+    "orchestra/testbench": "^10.0",
     "phpstan/phpstan": "^1.2",
     "yamadashy/phpstan-friendly-formatter": "^1.0.0"
   },
